Updated the README and cleaned up Dashboard
