+++ frontend/components/input-form.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
@@ -14,14 +15,17 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { getRoofArea, getRoofSegments } from "@/actions/google-solar";
+import { getRoofSegments } from "@/actions/google-solar";
 import { useLocationStore } from "@/stores/locations";
 import { useFormValueStore } from "@/stores/form-values";
 import { useParameterStore } from "@/stores/parameters";
 import { useEffect } from "react";
 import { calculateNumberOfPanels } from "@/lib/panel";
+import { FormError } from "@/components/form-error";
 
 export function InputForm() {
+  const [error, setError] = useState<string | undefined>("");
+
   const addLocation = useLocationStore((state) => state.addLocation);
   const formValues = useFormValueStore((state) => state);
   const setFormValues = useFormValueStore((state) => state.setValues);
@@ -41,11 +45,17 @@ export function InputForm() {
   }, [form, formValues]);
 
   async function onSubmit(values: z.infer<typeof LocationFormSchema>) {
-    const data = await getRoofArea(values.latitude, values.longitude);
-    const segments = await getRoofSegments(values.latitude, values.longitude);
-    console.log(segments);
+    setError("");
+    const { roofSegments, roofArea, error } = await getRoofSegments(
+      values.latitude,
+      values.longitude
+    );
+    if (error) {
+      setError(error);
+      return;
+    }
     const numberPanels = calculateNumberOfPanels(
-      segments,
+      roofSegments,
       panelParameters.length,
       panelParameters.width
     );
@@ -53,10 +63,10 @@ export function InputForm() {
       values.name,
       values.latitude,
       values.longitude,
-      data,
+      roofArea,
       numberPanels,
       true,
-      segments
+      roofSegments
     );
     form.reset();
     setFormValues(null, null);
@@ -112,6 +122,7 @@ export function InputForm() {
             </FormItem>
           )}
         />
+        <FormError message={error} />
         <Button type="submit" className="w-full">
           Add Location
         </Button>
